+++ esphome/components/ebus/binary_sensor/__init__.py
@@ -35,5 +35,4 @@ async def to_code(config):
     sensor_base_config(sens, config)
     cg.add(sens.set_response_read_mask(config[CONF_TELEGRAM][CONF_DECODE][CONF_MASK]))
 
-    cg.add(ebus.add_receiver(sens))
-    cg.add(ebus.add_sender(sens))
+    cg.add(ebus.add_sensor(sens))
