+++ esphome/components/ebus/sensor/ebus_sensor.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../ebus_sensor_base.h"
+#include "../ebus_component.h"
 #include "esphome/components/sensor/sensor.h"
 
 namespace esphome {
